requirements.txt
secp256k1
websocket-client
python-dotenv
bech32
websockets
fastapi
uvicorn
Jinja2
cryptography